disable poor man compute metadata cache
This caching paradigm is wrong. metadataCollector is instanciated only
once for the entire module.

Thus by not overriding computeMetadata the metadata of the first
response handled by the module was used for all subsequent events.

Clearly due to the implementation this had a very small impact overall,
but affected machineType